Remove folders and update import statement
