+++ .gitignore
@@ -11,7 +11,7 @@ package-lock.json
 
 # production
 /build
-/ui/dist
+/dist
 
 # misc
 .DS_Store
