+++ LICENSE
@@ -1,6 +1,6 @@
 BSD 3-Clause License
 
-Copyright (c) 2018, Eaton
+Copyright (c) 2018 - present, Eaton
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
