+++ package.json
@@ -1,14 +1,16 @@
 {
-    "name": "brightlayer-ui-colors",
-    "version": "0.0.0",
+    "name": "@brightlayer-ui/colors",
+    "version": "4.0.0-alpha.1",
     "description": "Brightlayer UI colors for Eaton applications",
-    "main": "index.ts",
+    "main": "./dist/commonjs/index.js",
+    "types": "./dist/commonjs/index.d.ts",
     "scripts": {
         "test": "./buildTest.sh",
-        "build": "cd ui && yarn build && cd ..",
+        "build": "yarn && tsc --module commonjs --outDir dist/commonjs && tsc --module ES2015 --outDir dist/es2015",
         "prettier": "prettier \"**/**.{ts,tsx,js,jsx,json,css,scss,html,md}\" --write",
-        "generate:licenses": "cd ./ui && npm-license-crawler -onlyDirectDependencies -json LICENSES.json && cd .. && npm-license-crawler -onlyDirectDependencies -json LICENSES.json",
-        "precommit": "yarn && yarn prettier && yarn generate:licenses && yarn test"
+        "precommit": "yarn && yarn prettier && yarn build && yarn test",
+        "publish:package": "set npm_config_yes=true && npx -p @brightlayer-ui/publish blui-publish",
+        "tag:package": "npx -p @brightlayer-ui/tag blui-tag"
     },
     "repository": {
         "type": "git",
@@ -22,7 +24,7 @@
         "Quality",
         "Lighting"
     ],
-    "author": "brightlayer-ui <brightlayer-ui@eaton.com>",
+    "author": "Brightlayer UI <brightlayer-ui@eaton.com>",
     "license": "BSD-3-Clause",
     "bugs": {
         "url": "https://github.com/etn-ccis/blui-colors/issues"
@@ -30,8 +32,17 @@
     "homepage": "https://github.com/etn-ccis/blui-colors#readme",
     "prettier": "@brightlayer-ui/prettier-config",
     "devDependencies": {
+        "typescript": "^5.0.3",
         "@brightlayer-ui/prettier-config": "^1.0.2",
-        "npm-license-crawler": "^0.2.1",
         "prettier": "^3.0.3"
-    }
+    },
+    "files": [
+        "package.json",
+        "README.md",
+        "LICENSE",
+        "CHANGELOG.md",
+        "dist/",
+        "branding-pallete.scss",
+        "palette.scss"
+    ]
 }
