+++ src/index.ts
@@ -1,0 +1,3 @@
+export * from './colors/palette';
+
+export { BrandingColors, chartColors as ChartColors } from './branding/palette';
