+++ src/types/colorType.ts
