+++ tsconfig.json
@@ -7,5 +7,5 @@
         "strict": true,
         "moduleResolution": "node"
     },
-    "include": ["src/*", "src/branding/*", "src/colors/*", "src/types/*"]
+    "include": ["src/*", "src/branding"]
 }
