--- ui/package.json
@@ -1,42 +1,0 @@
-{
-    "name": "@brightlayer-ui/colors",
-    "version": "4.0.0-alpha.0",
-    "description": "Brightlayer UI colors for Eaton applications",
-    "main": "./dist/commonjs/index.js",
-    "types": "./dist/commonjs/index.d.ts",
-    "scripts": {
-        "test": "echo \"Error: no test specified\" && exit 1",
-        "build": "yarn && tsc --module commonjs --outDir dist/commonjs && tsc --module ES2015 --outDir dist/es2015",
-        "publish:package": "set npm_config_yes=true && npx -p @brightlayer-ui/publish blui-publish",
-        "tag:package": "npx -p @brightlayer-ui/tag blui-tag"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/etn-ccis/blui-colors.git"
-    },
-    "keywords": [
-        "Brightlayer-UI",
-        "colors",
-        "Eaton",
-        "quality",
-        "lighting"
-    ],
-    "author": "Brightlayer UI <brightlayer-ui@eaton.com>",
-    "license": "BSD-3-Clause",
-    "bugs": {
-        "url": "https://github.com/etn-ccis/blui-colors/issues"
-    },
-    "homepage": "https://github.com/etn-ccis/blui-colors#readme",
-    "devDependencies": {
-        "typescript": "^5.0.3"
-    },
-    "files": [
-        "package.json",
-        "README.md",
-        "LICENSE",
-        "CHANGELOG.md",
-        "dist/",
-        "branding-pallete.scss",
-        "palette.scss"
-    ]
-}
