--- ui/src/branding/index.ts
@@ -1,21 +1,0 @@
-/**
- Copyright (c) 2018-present, Eaton
-
- All rights reserved.
-
- This code is licensed under the BSD-3 license found in the LICENSE file in the root directory of this source tree and at https://opensource.org/licenses/BSD-3-Clause.
- **/
-
-import { BrandingColors } from './palette';
-export const chartColors = [
-    BrandingColors.teal,
-    BrandingColors.citrus,
-    BrandingColors.pine,
-    BrandingColors.emerald,
-    BrandingColors.wine,
-    BrandingColors.crimson,
-    BrandingColors.sunset,
-    BrandingColors.rust,
-    BrandingColors.navy,
-    BrandingColors.sky,
-];
