--- ui/src/colors/index.ts
@@ -1,10 +1,0 @@
-/**
-Copyright (c) 2018-present, Eaton
- 
-All rights reserved.
- 
-This code is licensed under the BSD-3 license found in the LICENSE file in the root directory of this source tree and at https://opensource.org/licenses/BSD-3-Clause.
-**/
-
-export * from './palette';
-export * as BLUIColors from './palette';
