--- ui/src/index.ts
@@ -1,5 +1,0 @@
-export * from './colors/palette';
-export * as BLUIColors from './colors/palette';
-
-export * from './branding/palette';
-export * as Colors from './branding/palette';
